+++ frontend/app/tests/unit/store/settings/actions.spec.ts
@@ -1,8 +1,9 @@
 import { BigNumber } from '@rotki/common';
 import {
-  DARK_MODE_ENABLED,
   DARK_THEME,
-  LIGHT_THEME
+  LIGHT_THEME,
+  SELECTED_THEME,
+  Theme
 } from '@rotki/common/lib/settings';
 import {
   TimeFramePeriod,
@@ -73,7 +74,7 @@ describe('settings:actions', () => {
             [ITEMS_PER_PAGE]: 10,
             [AMOUNT_ROUNDING_MODE]: BigNumber.ROUND_UP,
             [VALUE_ROUNDING_MODE]: BigNumber.ROUND_DOWN,
-            [DARK_MODE_ENABLED]: false,
+            [SELECTED_THEME]: Theme.AUTO,
             [LIGHT_THEME]: LIGHT_COLORS,
             [DARK_THEME]: DARK_COLORS,
             [GRAPH_ZERO_BASED]: false,